+++ demisto_sdk/commands/common/tests/tools_test.py
@@ -73,6 +73,7 @@ from demisto_sdk.commands.common.tools import (
     get_latest_release_notes_text,
     get_marketplace_to_core_packs,
     get_pack_metadata,
+    get_pack_names_from_files,
     get_relative_path_from_packs_dir,
     get_release_note_entries,
     get_release_notes_file_path,
@@ -91,6 +92,7 @@ from demisto_sdk.commands.common.tools import (
     run_command_os,
     search_and_delete_from_conf,
     server_version_compare,
+    str2bool,
     string_to_bool,
     to_kebab_case,
 )
@@ -252,7 +254,7 @@ class TestGenericFunctions:
             )
         )
         assert "ü" in path.read_text(encoding="latin-1")
-        assert get_file(path, suffix) == {"text": SENTENCE_WITH_UMLAUTS}
+        assert get_file(path) == {"text": SENTENCE_WITH_UMLAUTS}
 
     @pytest.mark.parametrize(
         "file_name, prefix, result",
@@ -2675,6 +2677,36 @@ def test_get_from_version_error(mocker):
     assert str(e.value) == "yml file returned is not of type dict"
 
 
+@pytest.mark.parametrize(
+    "value, expected_output",
+    [
+        (None, False),
+        (True, True),
+        (False, False),
+        ("yes", True),
+        ("Yes", True),
+        ("YeS", True),
+        ("True", True),
+        ("t", True),
+        ("y", True),
+        ("Y", True),
+        ("1", True),
+        ("no", False),
+        ("No", False),
+        ("nO", False),
+        ("NO", False),
+        ("false", False),
+        ("False", False),
+        ("F", False),
+        ("n", False),
+        ("N", False),
+        ("0", False),
+    ],
+)
+def test_str2bool(value, expected_output):
+    assert str2bool(value) == expected_output
+
+
 PATH_1 = Path("1.yml")
 PATH_2 = Path("2.yml")
 
@@ -2911,3 +2943,43 @@ def test_is_content_item_dependent_in_conf(test_config, file_type, expected_resu
     """
     result = is_content_item_dependent_in_conf(test_config, file_type)
     assert result == expected_result
+
+
+@pytest.mark.parametrize(
+    "file_paths, skip_file_types, expected_packs",
+    [
+        (
+            [
+                "Packs/PackA/pack_metadata.json",
+                "Tests/scripts/infrastructure_tests/tests_data/collect_tests/R/Packs/PackB/pack_metadata.json",
+            ],
+            None,
+            {"PackA"},
+        ),
+        (
+            [("Packs/PackA/pack_metadata.json", "Packs/PackB/pack_metadata.json")],
+            None,
+            {"PackB"},
+        ),
+        (
+            ["Packs/PackA/pack_metadata.json", "Packs/PackB/ReleaseNotes/1_0_0.md"],
+            {FileType.RELEASE_NOTES},
+            {"PackA"},
+        ),
+    ],
+)
+def test_get_pack_names_from_files(file_paths, skip_file_types, expected_packs):
+    """
+    Given:
+        - Case A: Real packs paths and infra file paths.
+        - Case B: File paths in tuple.
+        - Case C: File paths and file types to skip.
+
+    When:
+        - Running get_pack_names_from_files.
+
+    Then:
+        - Ensure that the result is as expected.
+    """
+    packs_result = get_pack_names_from_files(file_paths, skip_file_types)
+    assert packs_result == expected_packs